pyproject.toml
[project]
name = "mcp-vods"
version = "0.1.8"
description = "MCP Server for Binge-watch / MoonTV / LunaTV. 一个用于追剧/追番的MCP服务器，为AI提供搜索影视播放地址的能力，并支持在小米等安卓电视/投影仪上直接播放"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["mcp", "llm", "tv", "movie", "video", "media", "binge"]
license = { text = "MIT" }
dependencies = [
    "aiohttp>=3.9.0",
    "fastmcp>=2.0.0",
    "PyYAML>=6.0.0",
    "cachetools>=6.2.0",
]

[project.urls]
Repository = "https://github.com/aahl/mcp-vods"

[project.scripts]
mcp-vods = "mcp_vods:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
